[Feature]: Adding Store to the database.
