+++ app/routes/user_routes.py
@@ -1,7 +1,7 @@
-from flask_socketio import SocketIO, emit
+from flask_socketio import SocketIO
 from flask import session, request, Blueprint, redirect, render_template, url_for
 
-from core.store_manager import STORE_REGISTRY
+from managers.store_manager.store_manager import STORE_REGISTRY
 from managers.user_manager import update_selected_stores
 from managers.user_manager.user_auth import update_password
 from managers.user_manager.user_manager import add_user, authenticate_user, get_user, update_username
