+++ managers/availability_manager/availability_update.py
@@ -2,7 +2,7 @@ import os
 import json
 import time
 from managers.redis_manager import redis_manager
-from core.store_manager import load_store_availability
+from managers.store_manager.store_manager import load_store_availability
 from managers.user_manager import get_user, get_user_directory, load_card_list, save_json
 from utility.logger import logger
