+++ managers/database_manager/__init__.py
@@ -1,3 +1,4 @@
-from .common_queries import add_user, get_user_by_username, update_username, update_password
+from .common_queries import add_user, get_user_by_username, update_username, update_password, get_store_metadata
 
-__all__ = ["add_user", "get_user_by_username", "update_username", "update_password"]
+__all__ = ["add_user", "get_user_by_username", "update_username", "update_password",
+           "get_store_metadata"]
