+++ managers/database_manager/tables.py
@@ -1,14 +1,29 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Text
+from sqlalchemy import Column, Integer, String, ForeignKey, Text, Table
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
 Base = declarative_base()
+
+# Association Table for User Store Preferences (Many-to-Many)
+user_store_preferences = Table(
+    "user_store_preferences",
+    Base.metadata,
+    Column("user_id", Integer, ForeignKey("users.id"), primary_key=True),
+    Column("store_id", Integer, ForeignKey("stores.id"), primary_key=True)
+)
+
+
+# Users Table (Updated)
 class User(Base):
     __tablename__ = "users"
+
     id = Column(Integer, primary_key=True, index=True)
-    username = Column(String, unique=True, index=True)
-    password = Column(String)
-    selected_stores = Column(Text)  # JSON list of stores
+    username = Column(String, unique=True, nullable=False)
+    password_hash = Column(String, nullable=False)
+
+    # Relationship to stores (User's selected stores)
+    selected_stores = relationship("Store", secondary=user_store_preferences, backref="users")
+
 
 class Card(Base):
     __tablename__ = "cards"
@@ -19,6 +34,7 @@ class Card(Base):
     finish = Column(String)
     image_url = Column(Text)  # Store the URL instead of the actual image
 
+
 class TrackedCard(Base):
     __tablename__ = "tracked_cards"
     id = Column(Integer, primary_key=True, index=True)
@@ -27,6 +43,7 @@ class TrackedCard(Base):
 
     user = relationship("User")
     card = relationship("Card")
+
 
 class Store(Base):
     __tablename__ = "stores"
