+++ managers/store_manager/store_manager.py
@@ -1,0 +1,11 @@
+import os
+import time
+
+
+def save_store_availability(store_name, card_name, listings):
+    """Save store availability data with a timestamp."""
+    pass
+
+def load_store_availability(store_name):
+    """Load all stored availability data for a given store."""
+    pass
