+++ managers/store_manager/stores/__init__.py
@@ -1,0 +1,5 @@
+from managers.store_manager.stores.authority_games_mesa_az import Authority_Games_Mesa_Arizona
+
+STORE_REGISTRY = {
+    "authority_games_mesa_az": Authority_Games_Mesa_Arizona()
+}
