+++ managers/user_manager/__init__.py
@@ -1,11 +1,11 @@
-from .user_manager import user_exists, get_user, add_new_user, update_username
-from .user_auth import add_user, authenticate_user
+from .user_manager import user_exists, get_user, add_user, update_username
+from .user_auth import authenticate_user
 from .user_preferences import update_selected_stores, get_selected_stores, load_user_config, save_user_config
 from .user_cards import load_card_list, save_card_list
 
 __all__ = [
-    "user_exists", "get_user", "add_new_user",
-    "add_user", "authenticate_user", "update_username",
+    "user_exists", "get_user", "add_user",
+    "authenticate_user", "update_username",
     "update_selected_stores", "get_selected_stores", "load_user_config", "save_user_config",
     "load_card_list", "save_card_list"
 ]
